Merge first final release 2024.4.7

+++ nrobo/util/common/__init__.py
@@ -1,11 +1,26 @@
+"""
+=====================CAUTION=======================
+DO NOT DELETE THIS FILE SINCE IT IS PART OF NROBO
+FRAMEWORK AND IT MAY CHANGE IN THE FUTURE UPGRADES
+OF NROBO FRAMEWORK. THUS, TO BE ABLE TO SAFELY UPGRADE
+TO LATEST NROBO VERSION, PLEASE DO NOT DELETE THIS
+FILE OR ALTER ITS LOCATION OR ALTER ITS CONTENT!!!
+===================================================
+
+
+"""
+
 import json
 import os.path as path
 import string
 import sys
-from random import random
+import random
+from pathlib import Path
 from time import time
+from typing import Union
 
 import yaml
+
 
 class Common:
     """
@@ -17,7 +32,7 @@ class Common:
     """
 
     @staticmethod
-    def read_file_as_string(file_path, encoding=None):
+    def read_file_as_string(file_path: Union[str, Path], encoding=None):
         """
         Read file as string
 
@@ -37,9 +52,10 @@ class Common:
                     return content
         except FileNotFoundError as file_not_found_error:
             print("No such file or directory found: " + file_path)
+            raise file_not_found_error
 
     @staticmethod
-    def write_text_to_file(file_path, content, encoding=None):
+    def write_text_to_file(file_path: Union[str, Path], content, encoding=None):
         """
         Write text to file
 
@@ -61,7 +77,7 @@ class Common:
             print("No such file or directory found: " + file_path)
 
     @staticmethod
-    def read_json(file_path):
+    def read_json(file_path: Union[str, Path]):
         """
         Read Json
 
@@ -79,7 +95,7 @@ class Common:
             print("No such file or directory found: " + file_path)
 
     @staticmethod
-    def write_json(file_path, dictionary):
+    def write_json(file_path: Union[str, Path], dictionary):
         """
         Write dictionary data to file
 
@@ -92,7 +108,7 @@ class Common:
             json.dump(dictionary, file, sort_keys=True, indent=4)
 
     @staticmethod
-    def is_file_exist(file_path):
+    def is_file_exist(file_path: Union[str, Path]):
         """
         Checks if given file exists or not.
 
@@ -102,15 +118,16 @@ class Common:
         return path.exists(file_path)
 
     @staticmethod
-    def read_yaml(file_path):
+    def read_yaml(file_path: Union[str, Path], / , *, fail_on_failure=True) -> Union[str, None]:
         """
         Read yaml file at given path
 
+        :param fail_on_failure: create file if TRUE though file is not present to read.
         :param file_path: Path of file
         :return: Content of yaml file -> dict()
         """
 
-        if not path.exists(file_path):
+        if not path.exists(file_path) and not fail_on_failure:
             """if file does not exist, then let's create it first"""
 
             with open(file_path, 'w') as file:
@@ -118,6 +135,10 @@ class Common:
 
                 # initialize file with empty dictionary
                 yaml.dump({}, file)
+
+        if not path.exists(file_path) and fail_on_failure:
+            """file does not exist"""
+            raise Exception(f"File {file_path} does not exist!")
         else:
             """Do Nothing as file exists"""
 
@@ -131,7 +152,7 @@ class Common:
             return data
 
     @staticmethod
-    def write_yaml(file_path, dictionary):
+    def write_yaml(file_path: Union[str, Path], dictionary):
         """
         Write dictionary data to given file path in yaml format
 
@@ -143,3 +164,20 @@ class Common:
         with open(file_path, 'w') as file:  # Open given file in write mode
             yaml.dump(dictionary, file)
 
+    @staticmethod
+    def generate_random_numbers(min, max):
+        """
+        Generate and return a random number in given range denoted by min and max
+
+        :param min:
+        :param max:
+        :return:
+        """
+        """
+        Returns a random string of given length
+
+        @Returns string a random string
+        """
+        random_number = random.randint(min, max)
+
+        return random_number
